+++ app-frontend/src/app/components/pages/project/analyses/index.js
@@ -1,11 +1,13 @@
 import analysesPage from './analyses';
 import comparePage from './compare';
 import visualizePage from './visualize';
+import quickeditPage from './quickedit';
 import settingsPages from './settings';
 
 export default [
     analysesPage,
     comparePage,
+    quickeditPage,
     visualizePage,
     ...settingsPages
 ];