+++ app-frontend/src/app/services/analysis/analysis.service.js
@@ -14,7 +14,7 @@ export default (app) => {
                     id: '@properties.id'
                 }, {
                     query: {
-                      method: 'GET',
+                        method: 'GET',
                         cache: false
                     },
                     get: {
@@ -107,7 +107,7 @@ export default (app) => {
                 let numAnalyses = page.count;
                 let requests = [firstRequest];
                 if (page.count > pageSize) {
-                    let requestMaker = function *(totalResults) {
+                    let requestMaker = function* (totalResults) {
                         let pageNum = 1;
                         while (pageNum * pageSize <= totalResults) {
                             let pageParams = {
@@ -141,23 +141,33 @@ export default (app) => {
         }
 
         getTemplate(id) {
-            return this.Template.get({id}).$promise;
+            return this.Template.get({
+                id
+            }).$promise;
         }
 
         deleteTemplate(id) {
-            return this.Template.delete({id}).$promise;
+            return this.Template.delete({
+                id
+            }).$promise;
         }
 
         getTemplateActions(id) {
-            return this.Template.actions({id}).$promise;
+            return this.Template.actions({
+                id
+            }).$promise;
         }
 
         deleteAnalysis(id) {
-            return this.Analysis.delete({id}).$promise;
+            return this.Analysis.delete({
+                id
+            }).$promise;
         }
 
         getAnalysis(id) {
-            return this.Analysis.get({id}).$promise;
+            return this.Analysis.get({
+                id
+            }).$promise;
         }
 
         createTemplate(template) {
@@ -199,7 +209,9 @@ export default (app) => {
 
         getNodeHistogram(analysis, nodeId) {
             return this.Analysis.histogram({
-                analysisId: analysis, node: nodeId, voidCache: true,
+                analysisId: analysis,
+                node: nodeId,
+                voidCache: true,
                 token: this.authService.token()
             }).$promise;
         }
@@ -223,7 +235,10 @@ export default (app) => {
             // Otherwise, update analysis in place
             // We're doing a depth-first graph traversal on the arguments here.
             // Start "before the beginning" of the arguments of the root operator.
-            let parents = [{node: analysis.executionParameters, argIdx: -1}];
+            let parents = [{
+                node: analysis.executionParameters,
+                argIdx: -1
+            }];
             if (analysis.executionParameters.id === targetId) {
                 Object.assign(analysis.executionParameters, assignment);
             }
@@ -242,7 +257,10 @@ export default (app) => {
                     let args = currChild.args || false;
                     // This child node is itself a parent. Push onto parents
                     if (args) {
-                        parents.push({node: currChild, argIdx: -1});
+                        parents.push({
+                            node: currChild,
+                            argIdx: -1
+                        });
                     }
                     // Edit it if it matches target. Usually this is leaf nodes, but not always
                     // (e.g. render definitions).
@@ -279,7 +297,10 @@ export default (app) => {
             }
             // We're doing a depth-first graph traversal on the arguments here.
             // Start "before the beginning" of the arguments of the root operator.
-            let parents = [{node: analysis.executionParameters, argIdx: -1}];
+            let parents = [{
+                node: analysis.executionParameters,
+                argIdx: -1
+            }];
             if (analysis.executionParameters.id === nodeId) {
                 return Object.assign({}, analysis.executionParameters.metadata);
             }
@@ -298,7 +319,10 @@ export default (app) => {
                     let args = currChild.args || false;
                     // This child node is itself a parent. Push onto parents
                     if (args) {
-                        parents.push({node: currChild, argIdx: -1});
+                        parents.push({
+                            node: currChild,
+                            argIdx: -1
+                        });
                     }
                     // Edit it if it matches target. Usually this is leaf nodes, but not always
                     // (e.g. render definitions).
@@ -332,7 +356,9 @@ export default (app) => {
             }
             nodes.forEach(n => {
                 // Only use projects, because constant nodes also have no apply.
-                if (!n.apply && (n.type === 'src' || n.type === 'projectSrc')) {
+                if (!n.apply && (n.type === 'src' ||
+                        n.type === 'projectSrc' ||
+                        n.type === 'layerSrc')) {
                     if (sourceIds.indexOf(n.id) < 0) {
                         sourceIds.push(n.id);
                         sources[n.id] = n;
@@ -371,28 +397,26 @@ export default (app) => {
         }
 
         getAnalysisActions(id) {
-            return this.Analysis.actions({id}).$promise;
+            return this.Analysis.actions({
+                id
+            }).$promise;
         }
 
         getAnalysisTileUrl(analysisId, params) {
             const token = this.authService.token();
-            const formattedParams = L.Util.getParamString(_.omitBy(Object.assign(
-                {
-                    token: this.authService.token(),
-                    tag: new Date().getTime()
-                }, params
-            ), (i) => !i));
+            const formattedParams = L.Util.getParamString(_.omitBy(Object.assign({
+                token: this.authService.token(),
+                tag: new Date().getTime()
+            }, params), (i) => !i));
             return `${this.tileServer}/tools/${analysisId}/{z}/{x}/{y}/${formattedParams}`;
         }
 
         getAnalysisTileUrlForProject(projectId, analysisId, params) {
             const token = this.authService.token();
-            const formattedParams = L.Util.getParamString(_.omitBy(Object.assign(
-                {
-                    token: params.mapToken ? null : this.authService.token(),
-                    tag: new Date().getTime()
-                }, params
-            ), (i) => !i));
+            const formattedParams = L.Util.getParamString(_.omitBy(Object.assign({
+                token: params.mapToken ? null : this.authService.token(),
+                tag: new Date().getTime()
+            }, params), (i) => !i));
             return `${this.tileServer}/${projectId}/analyses/${
                 analysisId}/{z}/{x}/{y}/${formattedParams}`;
         }